Extracting Data from XML
Extracting Data from XML
In this assignment you will write a Python program somewhat similar to http://www.pythonlearn.com/code/geoxml.py. 
The program will prompt for a URL, read the XML data from that URL using urllib and then parse and extract 
the comment counts from the XML data, compute the sum of the numbers in the file.